NetworkTV関連CtrlCmdをID指定に対応させる
LuaインタフェースのOpenNetworkTV()などに相当。
旧コマンドでは2つ以上の視聴用Viewアプリの起動や起動の確認、プロセスIDの取得ができないため。
都合が良いためSET_CH_INFOのspaceとchをint(int32_t)型に変更。intが負になるような有効なDWORD値はここでは使われず、
互換上の支障はない。また、この構造体とやり取りするCH_DATA4やTVTestのspaceやchはintなので、このほうが自然と思う。

+++ EpgTimerPlugIn/EpgTimerPlugIn/EpgTimerPlugIn.cpp
@@ -227,9 +227,7 @@ void CEpgTimerPlugIn::CtrlCmdCallbackInvoked()
 								ch = 0;
 							}
 						}else{
-							if( chInfo.Space == (int)val.space &&
-								chInfo.Channel == (int)val.ch )
-							{
+							if( chInfo.Space == val.space && chInfo.Channel == val.ch ){
 								if( sys->m_pApp->SetChannel(space, ch, val.SID) == true ){
 									resParam->param = CMD_SUCCESS;
 									AddDebugLog(L"TvTest:m_pApp->SetChannel true");
